requirements.txt
scrapy>=2.13.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
pymongo[srv]>=4.6.0
dnspython>=2.4.0
dateparser>=1.1.0
google-generativeai>=0.7.2
